+++ clients/rust/Cargo.toml
@@ -8,12 +8,10 @@ readme = "README.md"
 license-file = "../../LICENSE"
 
 [features]
-anchor = ["dep:anchor-lang"]
 test-sbf = []
 serde = ["dep:serde", "dep:serde_with"]
 
 [dependencies]
-anchor-lang = { version = "0.30.0", optional = true }
 borsh = "^0.10"
 num-derive = "^0.3"
 num-traits = "^0.2"
@@ -21,8 +19,3 @@ serde = { version = "^1.0", features = ["derive"], optional = true }
 serde_with = { version = "^3.0", optional = true }
 solana-program = { git = "https://github.com/2501babe/solana", branch = "hana-neostake-checkpoint" }
 thiserror = "^1.0"
-
-[dev-dependencies]
-assert_matches = "1.5.0"
-solana-program-test = { git = "https://github.com/2501babe/solana", branch = "hana-neostake-checkpoint" }
-solana-sdk = { git = "https://github.com/2501babe/solana", branch = "hana-neostake-checkpoint" }
