+++ interface/Cargo.toml
@@ -22,6 +22,7 @@ solana-instruction = { version="2.1.0", features=["bincode", "std"], git="https:
 solana-program = { version="2.1.0", git="https://github.com/anza-xyz/agave.git" }
 solana-program-error = { version="2.1.0", features=["serde"], git="https://github.com/anza-xyz/agave.git" }
 solana-pubkey = { version="2.1.0", features=["serde"], git="https://github.com/anza-xyz/agave.git" }
+solana-system-interface = { version="0.0.1", git="https://github.com/solana-program/system.git", branch="febo/system-instruction" }
 
 [dev-dependencies]
 assert_matches = "1.5.0"
